creada aplicacion users y endpoint GET users
