+++ site_app/voltix/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
+
 from . import views
 
 urlpatterns = [
-    path('usuarios/', views.get_all_users, name='get_all_users'),
+    path("", views.index, name="index"),
 ]
